Add charts redrawing
