+++ src/components/OrderBookChart/Chart/Models.tsx
@@ -1,0 +1,11 @@
+import {Order} from '../../../models';
+
+export interface ChartProps {
+  asks: Order[];
+  bids: Order[];
+  mid: string;
+}
+
+export interface OrdersProps {
+  orders: Order[];
+}
